+++ app/controllers/application_controller.rb
@@ -28,6 +28,10 @@ class ApplicationController < ActionController::Base
 
   private
 
+  def default_serializer_options
+    {root: false}
+  end
+
   def ensure_http_referer_is_set
     request.env['HTTP_REFERER'] ||= '/'
   end