+++ app/controllers/users_controller.rb
@@ -136,8 +136,11 @@ class UsersController < ApplicationController
   end
 
   def export
-    exporter = Diaspora::Exporter.new(Diaspora::Exporters::XML)
-    send_data exporter.execute(current_user), :filename => "#{current_user.username}_diaspora_data.xml", :type => :xml
+    if export = Diaspora::Exporter.new(current_user).execute
+      send_data export, filename: "#{current_user.username}_diaspora_data.json", type: :json
+    else
+      head :not_acceptable
+    end
   end
 
   def export_photos