+++ app/serializers/export/aspect_serializer.rb
@@ -1,0 +1,7 @@
+module Export
+  class AspectSerializer < ActiveModel::Serializer
+    attributes :name,
+               :contacts_visible,
+               :chat_enabled
+  end
+end